backend/requirements.txt
Flask==2.0.3
Flask_SQLAlchemy==2.5.1
Flask_Migrate==3.1.0
python-dotenv==0.19.2
google-cloud-speech==2.15.1
openai==0.27.0
textblob==0.17.1
Flask-Cors==3.0.10
